Fix issue where GOROOT was not picked up from env vars.
GOROOT and GOROOTForCmd are documented as:

	GOROOT, if specified, is made absolute (prefixed with the
	directory that the repository being built is checked out to)
	and is set as the GOROOT environment variable.

	GOROOTForCmd, if set, is used as the GOROOT env var when
	invoking the "go" tool.

However, in commit 8ad0bf4df4faaa689afb7447358064ff4e7e8bcb, it became
broken in that when config doesn't override GOROOT, GOROOT should come
from env vars, and hence GOROOTForCmd should be the same as normal
GOROOT. Instead, it was being set to an empty string, ignoring the
GOROOT in env vars. This was causing Travis test to fail because it was
looking for GOROOT in the wrong place (the right place was set in env
vars). /cc @sqs