vendor: Support go1.5 style vendoring
If we find a package which is under `vendor` directory, we record adjust it's
name similarly to how we do it for godeps. Unlike godeps wed do not adjust the
ImportPath since go1.5 vendoring solution does not do that (it does not GOPATH
munging).

Some repos pre go1.5 would use `vendor/src` style vendoring. If a package path
is under `vendor/src`, we do not do the go1.5 style vendoring.