graph: accept a list of source units as input to graph
This change will allow srclib-go to graph multiple packages together.
The benefit of this is that we can graph all source units of a repo
together, taking advantage of the caching in go/loader and gog packages,
to avoid doing duplicate work of loading and parsing the same dependencies
multiple times.

TODO: fix the global config setting in graph.go