depresolve: use newer tools/go/vcs package to resolve import paths
This changes the ResolveDep method to use the vcs.RepoRootForImportPath
method to resolve external package import paths, which is more stable
than the earlier gosrc.Get call. Additionally, this new call does not
incur a GitHub API request.

This change also makes it so that the Unit name of the resolved dep is
modified to have the base set to the root of the clone URL of the repo.
This ensures that the ref is resolved to the proper def, since the
def will not have a custom import path in its unit name when graphed by
srclib-go (since srclib-go does not have information about a repo's
custom import paths).

For eg, the golang.org/x/net/context.Context def will have the def key:
   Repo: github.com/golang/net
   Unit: github.com/golang/net/context
   UnitType: GoPackage
   Path: Context

Note how the Unit name is 'github.com/golang/net/context' and not
'golang.org/x/net/context'. This change ensures that refs to such defs
are always resolved correctly.