depresolve: Fix incorrect mapping for grpc import paths.
grpc imports paths do not follow the same pattern as the general
google.golang.org/... case.

This is a quick and easy solution, to add a special case for grpc. A
future investigation can see if it's viable to remove the special case
optimization, or if it's still worth having it for performance reasons.

Fixes #88.