depresolve: fixed bugs and added more special cases so cloneURLs are correct; added tests.
Fixed so that not all "google.golang.org" repos are treated the same,
instead three (of the four) have special cases. These are for efficiency
reasons, and because sourcegraph.com hosts their github.com mirrors.

Fixed bugs. Instead of doing a strings.Replace to get the ToRepoCloneURL,
the cloneURLs are just put in for special cases, for example "google.golang.org/grpc".

Tests are added to ensure correctness and prevent regressions.
The tests are marked with a build tag so that they are not
called every time this repo is built, as depresolve makes network calls that can be slow.

These changes make depresolve more accurate, which in turn helps
any project that wants a more accurate CloneURL resolution.