crude workaround for path collisions
This is to avoid a path collision between an interface method's name and a parameter with the same name of another method of the same interface. Tried long to find some less radical workaround, found none. Proper fix needs major rewrite of this code based on AST traversal. Feeling bad...